perf(client): remove unused entry data and defer settlement panels
