--- apps/game/src/ui/features/progression/index.ts
@@ -1,10 +1,0 @@
-// Progression Feature - Quests, onboarding, tutorials, hints
-// This feature handles player progression and learning systems
-
-// Quest System - Exports used externally
-export * from "./quests/quest-utils";
-
-// Tutorial & Hints System - Exports used externally
-export { ExplorationTable } from "./hints/exploration-table";
-export { HintModal, HintSection } from "./hints/hint-modal";
-export * from "./hints/utils";
