--- apps/game/src/ui/features/settlement/constants/index.ts
@@ -1,1 +1,0 @@
-export * from "./settlement-constants";
