+++ apps/game/src/ui/features/settlement/index.ts
@@ -1,16 +1,1 @@
-// Settlement Feature - Settlement management, building, production
-// This feature handles all settlement-related functionality
-
-// Settlement State & Utils - Exports used externally
-export * from "./constants";
-export * from "./utils";
-
-// Settlement Types & Utils - Exports used externally
-export type { SettlementLocation } from "./utils/settlement-types";
-export { getBanksLocations } from "./utils/settlement-utils";
-
-// Construction System - Exports used externally
-export { SelectPreviewBuildingMenu } from "./construction/select-preview-building";
-
-// Production System - Exports used externally
 export { ProductionModal } from "./production/production-modal";
