--- apps/game/src/ui/features/settlement/utils/index.ts
@@ -1,2 +1,0 @@
-export * from "./settlement-utils";
-export * from "./settlement-types";
